requirements.txt
Django==5.2.7
django_ratelimit==4.1.0
openpyxl==3.1.5
pandas==2.3.3
Pillow==12.0.0
